backend/app/requirements.txt
fastapi
sqlalchemy
pydantic
uvicorn
mysql-connector-python
python-dotenv
scikit-learn
torch
joblib
numpy
typing_extensions
